+++ test/libsolidity/smtCheckerTests/loops/while_loop_array_assignment_memory_storage.sol
@@ -25,5 +25,3 @@ contract LoopFor2 {
 // ====
 // SMTSolvers: z3
 // ----
-// Warning 4984: (237-242): CHC: Overflow (resulting value larger than 2**256 - 1) might happen here.
-// Warning 4984: (263-266): CHC: Overflow (resulting value larger than 2**256 - 1) might happen here.